Added markdown support for posts
Using gem Redcarpet. Custom renderer added for not adding <p> tags
to the title, currently just sat in the Post model, which isn't ideal.
The title is no longer renderer automatically on the post page, and must
be specified in the body - this allows for greater flexibility, and
avoids ugly prepending of markdown syntax.